requirements.txt
opentelemetry-distro 
opentelemetry-exporter-otlp
opentelemetry-sdk
pycoingecko